+++ src/app/[variants]/DesktopRouter.tsx
@@ -1,31 +1,11 @@
 'use client';
 
 import dynamic from 'next/dynamic';
-import { memo, useMemo } from 'react';
-import { RouterProvider } from 'react-router-dom';
 
-import BootErrorBoundary from '@/components/BootErrorBoundary';
 import Loading from '@/components/Loading/BrandTextLoading';
 import type { Locales } from '@/types/locale';
 
-import { createDesktopRouter } from './desktopRouter.config';
-
-interface ClientRouterProps {
-  locale: Locales;
-}
-
-const ClientRouter = memo<ClientRouterProps>(({ locale }) => {
-  const router = useMemo(() => createDesktopRouter(locale), [locale]);
-  return (
-    <BootErrorBoundary fallback={<Loading />}>
-      <RouterProvider router={router} />
-    </BootErrorBoundary>
-  );
-});
-
-ClientRouter.displayName = 'ClientRouter';
-
-const DesktopRouterClient = dynamic(() => Promise.resolve(ClientRouter), {
+const DesktopRouterClient = dynamic(() => import('./DesktopClientRouter'), {
   loading: () => <Loading />,
   ssr: false,
 });
